+++ ChangeLog.rst
@@ -51,6 +51,7 @@ Tik Manager Version History
 * **Fix.601** CSS fix / tab label font sizes fixed to 12px
 * **Fix.601** Maya playblast / fixed error caused by the mp4 conversion (arouse with v3.0.513)
 * **Fix.602** Check New version python3 bug fixed
+* **Fix.603** 3ds Max playblast filename issue prior to 2020 fixed
 
 2.5
 ---